% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bincov.R
\name{bincov_file}
\alias{bincov_file}
\title{Create a new \code{bincov_file} object.}
\usage{
bincov_file(x, medians)
}
\arguments{
\item{x}{Path to the binned coverage matrix file.}

\item{medians}{Named numeric vector of median coverage values for every
sample in the binned coverage matrix file.}
}
\description{
Create a new \code{bincov_file} object.
}
