% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdtable.R
\name{expand_gdtable}
\alias{expand_gdtable}
\title{Expand genomic disorder regions.}
\usage{
expand_gdtable(x, prop)
}
\arguments{
\item{x}{A \code{data.table} produced by \code{\link[=read_gdtable]{read_gdtable}}.}

\item{prop}{The fraction of each region by which to expand it.}
}
\value{
A \code{data.table} with columns \code{qstart} and \code{qend} with the expanded
start and end.
}
\description{
Expand each genomic disorder region by some fraction of its size.
}
\details{
The start of the expanded region will not be less than 1 and the end of the
region will not be greater than 5536870912.
}
