% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_tabix_ranges}
\alias{expand_tabix_ranges}
\title{Expand ranges by a fraction of their size.}
\usage{
expand_tabix_ranges(x, f)
}
\arguments{
\item{x}{\code{data.table} The ranges to expand. It must have at least the
columns \code{chr}, \code{start}, and \code{end}.}

\item{f}{\code{double} The fraction by which to expand the ranges.}
}
\value{
\code{data.table} The expanded ranges.
}
\description{
The ranges are expanded to a minimum start of 1 and a maximum end of
536,870,912 which are the limits of the tabix library.
}
