% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr_plotter.R
\name{gdplotter}
\alias{gdplotter}
\title{Create a new \code{gdplotter} object.}
\usage{
gdplotter(gd, bincov_mat, segdups, carriers)
}
\arguments{
\item{gd}{A list-like object that represents a GD region.
It must contain the following named components:
\itemize{
\item chr: Contig of the GD region
\item start_GRCh38: start of the GD region
\item end_GRCh38: end of the GD region
\item svtype: SV type of the GD (either 'DEL' or 'DUP')
}}

\item{bincov_mat}{A \code{bincov_mat} object over the GD region.}

\item{segdups}{A \code{\link[GenomicRanges:GRanges-class]{GRanges}} object of hg38
segmental duplications.}

\item{carriers}{A character vector of sample IDs of the GD carriers.}
}
\value{
A \code{gdplotter} object.
}
\description{
A \code{gdplotter} is an object that can plot a genomic disorder (GD) region.
}
