% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svtrio.R
\name{is_denovo}
\alias{is_denovo}
\title{Test if an SV is de novo}
\usage{
is_denovo(x)
}
\arguments{
\item{x}{An object to test for de novo status.}
}
\value{
Is the object predicted to represent a de novo event?
}
\description{
\code{is_denovo} is a generic, but it only really works for an
\code{\link{svtrio}} object. It uses some low effort heuristics to check
whether the PE/SR/RD evidence for an SV supports a de novo status. It is
only possible to confirm the de novo status of a variant if the evidence for
all members of the trio, i.e. child, mother, father, are available.
}
