% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bincov.R
\name{normalize_bincov}
\alias{normalize_bincov}
\title{Normalize coverage values from the bincov matrix.}
\usage{
normalize_bincov(x, medians)
}
\arguments{
\item{x}{\code{data.table} Binned coverage values.}

\item{medians}{\code{double} Genome-wide median coverage for each sample in \code{x}.}
}
\value{
\code{data.table} Normalized coverage values.
}
\description{
Normalization is done by dividing each sample's coverage values by the
genome-wide median coverage of the sample.
}
