% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe.R
\name{pe_file}
\alias{pe_file}
\title{Create a new \code{pe_file} object.}
\usage{
pe_file(x)
}
\arguments{
\item{x}{Path to the PE matrix file.}
}
\value{
A \code{pe_file} object.
}
\description{
Create a new \code{pe_file} object.
}
