% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_plotter.R
\name{pe_plotter}
\alias{pe_plotter}
\title{Create a new \code{pe_plotter} object.}
\usage{
pe_plotter(x)
}
\arguments{
\item{x}{A \code{\link[=query]{pe_mat}} object.}
}
\value{
A \code{pe_plotter} object.
}
\description{
A \code{pe_plotter} can create a visualization of discordant paired-end reads
alignments.
}
