% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denovo_plotter.R
\name{plot.denovo_plotter}
\alias{plot.denovo_plotter}
\title{Plot a \code{denovo_plotter} object.}
\usage{
\method{plot}{denovo_plotter}(x, y, ...)
}
\arguments{
\item{x}{A \code{\link{denovo_plotter}} object.}

\item{y}{Kept for compatibility with generic, but ignored here.}

\item{...}{Kept for compatibility with generic, but ignored here.}
}
\description{
Plot a \code{denovo_plotter} object.
}
