% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_plotter.R
\name{plot.pe_plotter}
\alias{plot.pe_plotter}
\title{Plot a \code{pe_plotter} object}
\usage{
\method{plot}{pe_plotter}(x, y, samples, main = "", col = "#000000", ...)
}
\arguments{
\item{x}{A \code{\link[=pe_plotter]{pe_plotter}} object.}

\item{y}{Ignored.}

\item{samples}{Samples to plot. The samples will be plotted from top to
bottom in the order they are given.}

\item{main}{Title for the plot.}

\item{col}{Color(s) to use for \code{samples}. Either a string or a character
vector of hexadecimal colors equal to the length of \code{samples}}

\item{...}{Ignored.}
}
\description{
The \code{plot()} method for \code{pe_plotter}.
}
\details{
The only additional parameter that is accepted is \code{col}, which should be a
string or a character vector equal to the length of \code{samples} giving the
colors to use for each sample.
}
\examples{
pe_path <- system.file("extdata", "example.PE.txt.gz", package = "gorilla", mustWork = TRUE)
pe <- pe_file(pe_path)
mat <- query(pe, "chr16", 28743149, 28745149)
plotter <- pe_plotter(mat)
plot(plotter, samples = "gorilla0003")
}
