% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr_plotter.R
\name{plot.rdr_plotter}
\alias{plot.rdr_plotter}
\title{Plot an \code{rdr_plotter}.}
\usage{
\method{plot}{rdr_plotter}(x, y = NULL, ...)
}
\arguments{
\item{x}{An \code{rdr_plotter} object.}

\item{y}{Kept for compatibility with the \code{plot} generic. Ignored here.}

\item{main}{An optional string to use as the title of the plot.}

\item{carrier}{Which carriers to plot, if \code{x} has carriers. If 0, no
carriers are plotting. If \code{NULL}, all carriers are plotted. If an integer
or character vector, it will be used as an index into the carriers vector.}
}
\description{
Plot an \code{rdr_plotter}.
}
