% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bincov.R, R/pe.R, R/query.R, R/rd.R, R/sr.R
\name{query.bincov_file}
\alias{query.bincov_file}
\alias{query.pe_file}
\alias{query}
\alias{query.rd_file}
\alias{query.sr_file}
\title{Query an object with a genomic region.}
\usage{
\method{query}{bincov_file}(x, contig, start, end)

\method{query}{pe_file}(x, contig, start, end)

query(x, contig, start, end)

\method{query}{rd_file}(x, contig, start, end)

\method{query}{sr_file}(x, contig, start, end)
}
\arguments{
\item{x}{An object.}

\item{contig}{The contig containing the region.}

\item{start}{The start of the region (1-start).}

\item{end}{The end of the region (inclusive).}
}
\value{
An object representing the queried region or \code{NULL} if the region
does not overlap any intervals in \code{x}.
}
\description{
\code{query()} retrieves some genomic region from \code{x}.
}
\details{
The method for \code{\link[=rd_file]{rd_file}} retrieves all the bins
overlapping a region from the RD matrix and returns an \code{rd_mat} object.

The method for \code{\link[=sr_file]{sr_file}} retrieves all split read alignments
overlapping a region from the SR matrix and returns an \code{sr_mat} object.

The method for \code{\link[=pe_file]{pe_file}} retrieves all aberrant paired-end
reads alignments overlapping a region from the PE matrix and returns a
\code{pe_mat} object. Only the upstream read is checked for overlap.
}
