% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bincov.R
\name{query_bincov}
\alias{query_bincov}
\title{Query ranges from a bincov matrix.}
\usage{
query_bincov(con, ranges, header = NULL, select = NULL)
}
\arguments{
\item{con}{\code{TabixFile} Connection object to bincov matrix.}

\item{ranges}{\code{GRanges} Genomic ranges to use as the query.}

\item{header}{\code{character} The parsed header line from the bincov matrix.}

\item{select}{\code{integer} Which columns to keep.}
}
\value{
\code{list} A list of \code{data.table}s, one for each range in \code{ranges}. If
a range does not overlap any intervals in the bincov matrix, the element
corresponding to that range will be \code{NULL}.
}
\description{
Query ranges from a bincov matrix.
}
