% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd.R
\name{rd_file}
\alias{rd_file}
\title{Create a new \code{rd_file} object.}
\usage{
rd_file(x, medians)
}
\arguments{
\item{x}{Path to the RD matrix file.}

\item{medians}{Named numeric vector of median coverage values of samples in
the RD matrix file.}
}
\value{
A \code{rd_file} object.
}
\description{
Create a new \code{rd_file} object.
}
