% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_plotter.R
\name{rd_plotter}
\alias{rd_plotter}
\title{Create a new \code{rd_plotter} object}
\usage{
rd_plotter(x, median_region = NULL)
}
\arguments{
\item{x}{A \code{\link[=query]{rd_mat}} object.}

\item{median_region}{A list with components \code{contig}, \code{start}, and \code{end}
to indicate that the median normalized read depth of that region for each
sample should be included in the plot. If \code{NULL}, the medians will not be
included.}
}
\value{
A \code{rd_plotter} object.
}
\description{
A \code{rd_plotter} can create a visualization of read depth.
}
\examples{
rd_path <- system.file(
    "extdata",
    "example.RD.txt.gz",
    package = "gorilla",
    mustWork = TRUE
)
medians_path <- system.file(
    "extdata",
    "example_medianCov.txt",
    package = "gorilla",
    mustWork = TRUE
)

rd <- rd_file(rd_path, medians_path)
mat <- query(rd, "chr16", 28743149, 28745149)
plotter <- rd_plotter(mat)
}
