% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr_plotter.R
\name{rdr_plotter}
\alias{rdr_plotter}
\title{Create a new \code{rdr_plotter} object.}
\usage{
rdr_plotter(x, segdups)
}
\arguments{
\item{x}{A \code{bincov_mat} object.}

\item{segdups}{A \code{\link[GenomicRanges:GRanges-class]{GRanges}} object with ranges for
segmental duplications.}
}
\value{
A \code{rdr_plotter} object.
}
\description{
An \code{rdr_plotter} object represents a region of a binned coverage matrix to
visualize.
}
