% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr_plotter.R
\name{rdr_plotter_set_carriers}
\alias{rdr_plotter_set_carriers}
\title{Set the carriers for an \code{rdr_plotter}.}
\usage{
rdr_plotter_set_carriers(x, carriers)
}
\arguments{
\item{x}{An \code{rdr_plotter} object.}

\item{carriers}{A named character vector of which the values are colors to
use for plotting the carriers and the names are the IDs of the carriers.}
}
\value{
An \code{rdr_plotter} object.
}
\description{
Set the carriers for an \code{rdr_plotter}.
}
