% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdr_plotter.R
\name{rdr_plotter_set_pad}
\alias{rdr_plotter_set_pad}
\title{Set the padding for an \code{rdr_plotter}.}
\usage{
rdr_plotter_set_pad(x, left = NULL, right = NULL)
}
\arguments{
\item{x}{An \code{rdr_plotter} object.}

\item{left}{The genomic coordinate of the left edge of the focal region.}

\item{right}{The genomic coordinate of the right egde of the focal region.}
}
\value{
An \code{rdr_plotter} object.
}
\description{
The padding represents the genomic region outside the region of interest.
This function does not increase the size of the region that can be plotted
by \code{x}. It merely indicates which part of the existing region should be
considered the focal region.
}
