% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bincov.R
\name{read_bincov_header}
\alias{read_bincov_header}
\title{Retrieve the header of a bincov matrix.}
\usage{
read_bincov_header(con)
}
\arguments{
\item{con}{\code{TabixFile} Connection object to bincov matrix.}
}
\value{
\code{character} Header of the matrix.
}
\description{
Retrieve the header of a bincov matrix.
}
