% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdtable.R
\name{read_gdtable}
\alias{read_gdtable}
\title{Read genomic disorders table file.}
\usage{
read_gdtable(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A \code{data.table}.
}
\description{
Reads a TSV file describing genomic disorder regions.
}
\details{
The file must have the following columns:
\enumerate{
\item chr: contig of the region
\item start_GRCh38: start of the region
\item end_GRCh38: end of the region
\item GD_ID: ID of the region
\item svtype: either 'DEL' or 'DUP'
\item NAHR: either 'yes' or 'no' indicating NAHR-mediated or not
\item terminal: one of 'p', 'q', or 'no' indicating the p-arm, q-arm, or
non-terminal region
\item cluster: either a cluster ID or an empty string
The file must have column headers as given.
}
}
