% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medians.R
\name{read_medians_file}
\alias{read_medians_file}
\title{Read coverage medians file into a named vector.}
\usage{
read_medians_file(path)
}
\arguments{
\item{path}{\code{character(1)} Path to the coverage medians file.}
}
\value{
\code{double} The coverage medians. The names of the vector are the
sample IDs.
}
\description{
Read coverage medians file into a named vector.
}
