% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segdups.R
\name{read_segdups}
\alias{read_segdups}
\title{Read a BEDX file with segmental duplication regions.}
\usage{
read_segdups(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A \code{GRanges} object.
}
\description{
The first three columns must be contig, start, and end in 0-start, open-end
coordinates.
}
