% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr.R
\name{sr_file}
\alias{sr_file}
\title{Create a new \code{sr_file} object.}
\usage{
sr_file(x)
}
\arguments{
\item{x}{Path to the SR matrix file.}
}
\value{
A \code{sr_file} object.
}
\description{
Create a new \code{sr_file} object.
}
