% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_plotter.R
\name{sr_plotter}
\alias{sr_plotter}
\title{Create a new \code{sr_plotter} object.}
\usage{
sr_plotter(x)
}
\arguments{
\item{x}{A \code{\link[=query]{sr_mat}} object.}
}
\value{
A \code{sr_plotter} object.
}
\description{
A \code{sr_plotter} can create a visualization of split read alignments.
}
