% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe.R, R/rd.R, R/sr.R, R/subset_samples.R,
%   R/svevidence.R
\name{subset_samples.pe_mat}
\alias{subset_samples.pe_mat}
\alias{subset_samples.rd_mat}
\alias{subset_samples.sr_mat}
\alias{subset_samples}
\alias{subset_samples.svevidence}
\title{Subset an object to some samples}
\usage{
\method{subset_samples}{pe_mat}(x, samples)

\method{subset_samples}{rd_mat}(x, samples)

\method{subset_samples}{sr_mat}(x, samples)

subset_samples(x, samples)

\method{subset_samples}{svevidence}(x, samples)
}
\arguments{
\item{x}{Object to subset.}

\item{samples}{String identifiers for the samples to keep.}
}
\value{
\code{x} subset to \code{samples}.
}
\description{
\code{subset_samples()} subsets \code{x} to the samples in \code{samples}.
}
\details{
Here, \code{samples} refers to entities in an experiment rather than the a random
subset. Methods should ignore samples in \code{samples} that are not present in
\code{x}.

There are methods for \code{pe_mat}, \code{sr_mat}, \code{rd_mat}, and \code{svevidence}.
}
\examples{
pe_path <- system.file("extdata", "example.PE.txt.gz", package = "gorilla", mustWork = TRUE)
pe <- pe_file(pe_path)
mat <- query(pe, "chr16", 28743149, 28747149)
mat_sub <- subset_samples(mat, c("gorilla0001", "gorilla0002"))
}
