% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svevidence.R
\name{supports_sv}
\alias{supports_sv}
\title{Does an SV have supporting evidence}
\usage{
supports_sv(x, sample_id)
}
\arguments{
\item{x}{Object that can be used to check SV support.}

\item{sample_id}{ID of sample to check.}
}
\value{
Does the SV have supporting evidence.
}
\description{
Check the evidence for an SV to see if the call is supported.
}
\details{
The method for \code{\link{svevidence}} will check if an SV is supported by PE, SR, or
RD evidence.
}
