% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svevidence.R
\name{svevidence}
\alias{svevidence}
\title{Create a new \code{svevidence} object.}
\usage{
svevidence(contig, start, end, pe, sr, rd, pad = 0)
}
\arguments{
\item{contig}{The contig containing the region.}

\item{start}{The start of the region (1-start).}

\item{end}{The end of the region (inclusive).}

\item{pe}{A \code{\link{pe_file}} object for the discordant read pairs evidence.}

\item{sr}{A \code{\link{sr_file}} object for the split read evidence.}

\item{rd}{A \code{\link{rd_file}} object for the read depth evidence.}

\item{pad}{Fraction of region to add as padding. The minimum padded start is
1 and the maximum padded end is 536870912.}
}
\value{
A \code{svevidence} object.
}
\description{
This class represents SV evidence over a genomic region. It is assumed that
the evidence sources all have the same set of samples.
}
