% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe.R, R/rd.R, R/sr.R, R/svevidence.R
\name{pe_file}
\alias{pe_file}
\alias{rd_file}
\alias{sr_file}
\alias{svevidencefiles}
\title{Structural variant evidence files}
\usage{
pe_file(path)

rd_file(path, medians)

sr_file(path)
}
\arguments{
\item{path}{Path to the evidence file.}

\item{medians}{Named numeric vector of median coverage values of samples.}
}
\value{
An object for the evidence type.
}
\description{
Classes to handle structural variant evidence files used in the GATK-SV
pipeline.
}
\details{
The supported files are almost identical to the ones generated by the GATK
tool
\href{https://gatk.broadinstitute.org/hc/en-us/articles/35967530268443-CollectSVEvidence-BETA}{CollectSVEvidence}
with the main difference being that the files referenced here combine
evidence from multiple samples into a single file. All files must be
compressed with bgzip and indexed with tabix. Each file/evidence type has a
wrapper class that is used to open the file and extract the evidence for a
genomic region. Each construction function returns an object of a class with
the same name as the constructor.
}
\section{Paired-end Reads}{
This file contains discordant read pairs and the format is the same as the
single sample version, though the "sample name" field may contain multiple
samples. An object for this file type can be constructed with the
\code{pe_file()} function.
}

\section{Split Reads}{
This file contains split read clippings and the format is the same as the
single sample version, though the "sample name" field may contain multiple
samples. An object for this file type can be constructed with the
\code{sr_file()} function.
}

\section{Read Depth}{
This file contains binned read depth. The file contains three columns
indentifying the interval and then columns, one per sample, of the read
depth of reads that start in each interval. There is also a header line that
starts with "#Chr", "Start", and "End" in the first three columns and then
the sample name of each sample in the file. An object for this file type can
be constructed with the \code{rd_file()}. The read depth file class additionally
requires a information on the median read depth of each sample. See
\code{\link[=read_median_coverages]{read_median_coverages()}}.
}

\examples{
# PE evidence

tf_pe_src <- tempfile()
tf_pe_dest <- tempfile(fileext = ".gz")
src_conn <- file(tf_pe_src, open = "wt")
cat("chr14\t10000\t+\tchr14\t20000\t+\tabc0000\n", file = tf_pe_src)
cat("chr14\t11000\t+\tchr14\t15000\t+\tabc0000\n", file = tf_pe_src)
cat("chr14\t11000\t+\tchr14\t16000\t+\tabc0001\n", file = tf_pe_src)
close(src_conn)
Rsamtools::bgzip(tf_pe_src, tf_pe_dest)
Rsamtools::indexTabix(tf_pe_dest, seq = 1, start = 2, end = 2)
pe <- pe_file(tf_pe_dest)

# RD evidence

tf_medians <- tempfile()
writeLines(c("abc0000\tabc0001", "30\t30"), tf_medians)
tf_rd_src <- tempfile()
tf_rd_dest <- tempfile(fileext = ".gz")
src_conn <- file(tf_rd_src, open = "wt")
cat("#Chr\tStart\tEnd\tabc0000\tabc0001\n", file = src_conn)
cat("chr5\t1000000\t1000100\t40\t41\n", file = src_conn)
cat("chr5\t1000100\t1000200\t39\t41\n", file = src_conn)
cat("chr5\t1000200\t1000300\t30\t35\n", file = src_conn)
close(src_conn)

Rsamtools::bgzip(tf_rd_src, tf_rd_dest)
Rsamtools::indexTabix(tf_rd_dest, seq = 1, start = 2, end = 3, comment = "#", zeroBased = TRUE)
rd_medians <- read_median_coverages(tf_medians)
rd <- rd_file(tf_rd_dest, rd_medians)

}
