% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denovo_plotter.R
\name{denovo_plotter}
\alias{denovo_plotter}
\title{Create a new \code{denovo_plotter} object.}
\usage{
denovo_plotter(x)
}
\arguments{
\item{x}{A \code{\link{svtrio}} object.}
}
\value{
A \code{denovo_plotter} object.
}
\description{
A \code{denovo_plotter} can create a visualization of de novo SV evidence for a
child, father, mother trio.
}
