% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forage.R
\name{forage}
\alias{forage}
\title{Forage for genomic disorders.}
\usage{
forage(
  gds_path,
  sds_path,
  bincov_path,
  medians_path,
  samples_path,
  sex_ploidy_path,
  outdir_path,
  min_shift = 0.3,
  pad = 0.5,
  max_calls_per_sample = 10,
  outliers_log_path = NULL
)
}
\arguments{
\item{gds_path}{Path to the genomic disorders table file. See
\code{\link[=read_gdtable]{read_gdtable}}}

\item{sds_path}{Path to a BEDX file of hg38 segmental duplication regions.}

\item{bincov_path}{Path to the binned coverage matrix.}

\item{medians_path}{Path to the median coverages file.}

\item{samples_path}{Path to a list of samples, one-per-line, to process.}

\item{sex_ploidy_path}{Path to a ploidy table. The format is a tab-delimited
file with the following columns:
\enumerate{
\item sample ID
\item chrX ploidy with 0 = unknown, 1 = male, 2 = female
The file must not have column headers.
}}

\item{outdir_path}{Path the output directory. It will be created if it
doesn't exist and any files that match the name of a plot created by the
workflow will be silently overwritten.}

\item{min_shift}{Minimum amount by which the read-depth ratio of a sample
must be shifted from expectation to be considered a GD carrier.}

\item{pad}{Proportion of the GD region that should be included on either
side when plotting.}

\item{max_calls_per_sample}{Maximum number of GD calls a sample can have
before it is marked as an outlier and all of its calls are treated as false.}

\item{outliers_log_path}{Path to a file where the IDs of any outlier samples
will be written. If \code{NULL}, outlier samples are not recorded.}
}
\description{
A convenience function to run the genomic disorder calling workflow on a set
of inputs.
}
