% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_plotter.R
\name{plot.pe_plotter}
\alias{plot.pe_plotter}
\title{Plot a \code{pe_plotter} object.}
\usage{
\method{plot}{pe_plotter}(x, y, samples, ...)
}
\arguments{
\item{x}{A \code{\link{pe_plotter}} object.}

\item{y}{Kept for compatability with generic, but ignored here.}

\item{samples}{Samples to plot. The samples will be plotted from top to
bottom in the order they are given.}

\item{...}{Other plotting parameters. Currently the only accepted parameter
is \code{col} which should either be a character vector of length 1 or length
equal to \code{samples} giving the colors to use for each sample.}
}
\description{
Plot a \code{pe_plotter} object.
}
