% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_plotter.R
\name{plot.rd_plotter}
\alias{plot.rd_plotter}
\title{Plot a \code{rd_plotter} object.}
\usage{
\method{plot}{rd_plotter}(x, y, samples, bg_samples = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{rd_plotter}} object.}

\item{y}{Kept for compatability with generic, but ignored here.}

\item{samples}{Samples to plot.}

\item{bg_samples}{Samples to use as background samples. If \code{NULL},
background samples are not included.}

\item{...}{Other plotting parameters. Currently the only accepted parameter
is \code{col} which should either be a character vector of length 1 or length
equal to \code{samples} giving the colors to use for each sample.}
}
\description{
Plot a \code{rd_plotter} object.
}
