% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bincov.R
\name{query}
\alias{query}
\alias{query.bincov_file}
\title{Query an object with a genomic region.}
\usage{
query(x, contig, start, end)

\method{query}{bincov_file}(x, contig, start, end)
}
\arguments{
\item{x}{An object.}

\item{contig}{The contig containing the region.}

\item{start}{The start of the region (1-start).}

\item{end}{The end of the region (inclusive).}
}
\value{
An object representing the queried region.
}
\description{
\code{query()} retrieves some genomic region from \code{x}.
}
\details{
The method for \code{\link[=bincov_file]{bincov_file}} retrieves all the bins
overlapping the requested region from the binned coverage matrix and returns
a \code{bincov_mat} object or \code{NULL} if the matrix does not have any overlappng
bins.
}
