% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_plotter.R
\name{rd_plotter}
\alias{rd_plotter}
\title{Create a new \code{rd_plotter} object.}
\usage{
rd_plotter(x)
}
\arguments{
\item{x}{A \code{\link[=query]{rd_mat}} object.}
}
\value{
A \code{rd_plotter} object.
}
\description{
A \code{rd_plotter} can create a visualization of read depth.
}
