% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdtable.R
\name{read_gdtable}
\alias{read_gdtable}
\title{Read a genomic disorders table file}
\usage{
read_gdtable(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A \code{\link[data.table:data.table]{data.table}}.
}
\description{
Read a TSV file describing genomic disorder regions.
}
\details{
The file must have the following columns:
\enumerate{
\item chr: contig of the region
\item start_GRCh38: start of the region
\item end_GRCh38: end of the region
\item GD_ID: ID of the region
\item svtype: either 'DEL' or 'DUP'
\item NAHR: either 'yes' or 'no' indicating NAHR-mediated or not
\item terminal: one of 'p', 'q', or 'no' indicating the p-arm, q-arm, or
non-terminal region
\item cluster: either a cluster ID or an empty string
The file must have column headers as given.
}
}
\examples{
tmp <- tempfile()
fp <- file(tmp, open = "wt")
writeLines(
    "chr\tstart_GRCh38\tend_GRCh38\tGD_ID\tsvtype\tNAHR\tterminal\tcluster",
    fp
)
writeLines("chr1\t898703\t6229913\tGD_0000\tDUP\tno\tp\t", fp)
writeLines("chr15\t31727418\t32153204\tGD_0001\tDEL\tyes\tno\tcluster00", fp)
close(fp)
gds <- read_gdtable(tmp)
}
