% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_coverages.R
\name{read_median_coverages}
\alias{read_median_coverages}
\title{Read the median coverages file.}
\usage{
read_median_coverages(path)
}
\arguments{
\item{path}{Path to the file.}
}
\value{
A named vector in which the names are the sample IDs and the values
are the median coverages.
}
\description{
Read the median coverages file.
}
