% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_plotter.R
\name{sr_plotter}
\alias{sr_plotter}
\title{Create a new \code{sr_plotter} object}
\usage{
sr_plotter(x)
}
\arguments{
\item{x}{A \code{\link[=query]{sr_mat}} object.}
}
\value{
A \code{sr_plotter} object.
}
\description{
A \code{sr_plotter} can create a visualization of split read alignments.
}
\examples{
sr_path <- system.file("extdata", "example.SR.txt.gz", package = "gorilla", mustWork = TRUE)
sr <- sr_file(sr_path)
mat <- query(sr, "chr16", 28743149, 28745149)
plotter <- sr_plotter(mat)
}
