% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svtrio.R
\name{svtrio}
\alias{svtrio}
\title{Create a new \code{svtrio} object.}
\usage{
svtrio(evidence, child, father, mother)
}
\arguments{
\item{evidence}{A \code{sveidence} object.}

\item{child}{The child sample.}

\item{father}{The father sample.}

\item{mother}{The mother sample.}
}
\value{
A \code{svtrio} object.
}
\description{
The \code{svtrio} object represents SV evidence over a genomic region for a
child, father, mother trio of samples.
}
